Add command for migrating brokers
There’s currently no easy way to use the bundled scripts to migrate
ownership between an old broker and a new one. The approach taken here
is to simply iterate through all topic-partition combinations and replace
the old broker id with the new broker id, and then execute a reassign
partitions command, which should make the new broker a follower to all
the topic-partition combinations that the old broker is responsible for,
leader or follower wise.

There’ll be additional commands for balancing a cluster, performing
preferred replica leader assignment and whatever else we can come up
with.