Add command for (attempting) to balance a cluster
Currently doesn’t account for previous ownership at all, thus it can
potentially move all topic-partitions to new leaders rather than trying
to be smart about it. We’ll try this out and see how it works and then
we can decide whether we want / need to have something more
sophisticated. I figured that it would be nice to start off using the
classes and whatnot that is already bundled with Kafka.