Attempt to make BalancePlan#generate a bit more deterministic
Noticed that the `#assign_replicas_to_brokers` randomises which broker
it starts from, which will essentially generate a new assignment even
though the topology hasn’t changed. In an attempt to circumvent this we
keep track of which broker index we used for the previous partition and
use the next broker for the next partition, etc, etc.

Perhaps we should have a `—randomize` flag to the `balance` command for
times when we actually just want to shuffle everything around, but I’ll
think about that.