Add ScalaEnumerable
Got a bit tired of not being able to use the nice Enumerable methods
that are available in Ruby, so this is a really simple wrapper.

Small test of this in the ZookeeperClient class, and also snuck in a
change regarding the resulting Map from the private `#merge` function,
mostly just to save a couple of characters.