Improve handling of overflow json file
If there’s a previous overflow present, and the user selects to go
ahead with a newly generated reassignment plan, the old overflow file
is removed _AFTER_ the new reassignment has been put in motion.

In the case of the new reassignment plan also causing an overflow, this
data will overwrite the previous one without asking the user, as this is
what should be expected.