Add integration tests for using overflow data
Realised that ZkUtils#read_data returns a tuple and not just a list of
child znodes, so that’s a thing.
Also noticed that $stdin#reopen doesn’t work in quite the same way on
JRuby as it does on MRI Ruby, which is a bit sad. So now I’ ve resorted
to stubbing Readline.readline calls, which I’m not proud of, but it
works for now.