Add command for showing progress of a reassignment
A “reassignment” is the term used when referring to the process of moving
partitions between brokers, be it leadership or just for replication
purposes.
Whenever a `balance`, `migrate` or `decommission` command is executed it
will generate a “reassignment plan”, which is written to a specific path
in ZooKeeper, which will kick-start the process of moving partitions in
the cluster. Ktl will also write the same JSON to a “state” prefix in
ZooKeeper, so it’s possible to see exactly which partitions are left to
move.

The terminology is a bit messy right now, and so is the organization of
commands. Not sure exactly what would be the most optimal solution, so
this'll do for now.