Fix faulty progress logic
Well, uhm, this is sort of embarrassing. The previous logic was really
whack in retrospect, which was also reflected in the tests. Sort of got
it backwards when I was thinking about how to “measure” progress of a
reassignment. As the JSON that’s currently in the “reassign_partitions”
path in ZK holds the partitions that are left to reassign.