Rename `balance` command to `shuffle`
The main motivation for this change is because the current ‘balancing’
doesn’t actually work too well if you have more than three (3) brokers,
as it tends to favours the first three (3) brokers, so we end up with a
uneven distribution of partitions amongst the brokers.

The original thought was to minimize moving partitions between brokers,
but the execution wasn’t up-to par, so for now I’m going to rename the
operation to “shuffle”, as it’s what it does, it’ll shuffle all partitions
(that match the given regexp) over the cluster.

Perhaps we could add a “balance” command that looks at the number of
messages currently present for each topic, but that’ll be a later
problem.