Use an embedded Kafka broker in integration tests
The previous setup shell'ed out and started one or more brokers, but
the thing was kind of messy since there's already an embedded ZooKeeper
process running.

Perhaps it would be beneficial to start ZooKeeper and Kafka separately
from the integration tests, so to test different versions of Kafka, but
right now it's just too messy as too many of the integration tests
would have to be rewritten as some of them are engineered for specific
scenarios that might be difficult to reliably reproduce without a lot
of work (and I don't currently have time for that).